@@ -1,7 +1,7 @@
 [package]
 name = "prover"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
 publish = false
 
 [dependencies]
