@@ -4,13 +4,12 @@
 #![allow(clippy::missing_safety_doc, clippy::too_many_arguments)]
 
 pub mod binary;
-/// cbindgen:ignore
-pub mod console;
 mod host;
 pub mod machine;
 /// cbindgen:ignore
 mod memory;
 mod merkle;
+/// cbindgen:ignore
 pub mod programs;
 mod reinterpret;
 pub mod utils;
